+++ apps/admin/src/app.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "@tryghost/admin-x-framework";
 import { useCurrentUser } from "@tryghost/admin-x-framework/api/current-user";
 import { EmberProvider, EmberFallback, EmberRoot } from "./ember-bridge";
-import { AdminLayout } from "./layout/AdminLayout";
+import { AdminLayout } from "./layout/admin-layout";
 import { useEmberAuthSync, useEmberDataSync } from "./ember-bridge";
 
 function App() {
