+++ apps/admin/src/layout/app-sidebar/icons/network-icon.tsx
